legacy/junas/backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[project]
name = "junas"
version = "1.0.0"
description = "Junas — unified legal AI platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",
  "pydantic-settings>=2.0",
  "asyncpg>=0.29",
  "sqlalchemy[asyncio]>=2.0",
  "alembic>=1.13",
  "elasticsearch[async]>=8.12",
  "qdrant-client>=1.8",
  "redis>=5.0",
  "celery>=5.3",
  "prefect>=2.14",
  "transformers>=4.38",
  "torch>=2.2",
  "sentence-transformers>=2.5",
  "datasets>=2.18",
  "seqeval>=1.2",
  "jieba>=0.42",
  "rank-bm25>=0.2",
  "scikit-learn>=1.4",
  "httpx>=0.27",
  "python-docx>=1.1",
  "pdfplumber>=0.11",
  "beautifulsoup4>=4.12",
  "lxml>=5.1",
  "python-multipart>=0.0.9",
  "sse-starlette>=2.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "ruff>=0.3",
  "mypy>=1.9"
]

[tool.hatch.build.targets.wheel]
packages = ["api", "data", "ml"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
